算法考核编写1.0

+++ src/main/java/cn/staitech/anno/domain/AlgorithmJson.java
@@ -69,7 +69,7 @@ public class AlgorithmJson implements Serializable {
     /**
      * 更新者
      */
-    private Integer updateBy;
+    private Long updateBy;
 
     /**
      * 更新时间
